- Updated number formatting to avoid scientific notation for small decimals (e.g. 0.00000035 being shows as 3.5e-7) - Updated BottomNavigationView to have icons - Updated OpenOrders page to show 200 orders instead of 20, and it now starts in the middle of the RecyclerView instead of at the top

+++ app/src/main/java/com/treehouse/gdax/Data/MatchOrder.kt
@@ -12,7 +12,7 @@ data class MatchOrder (
     @ColumnInfo var maker_order_id: String = "",
     @ColumnInfo var taker_order_id: String = "",
     @ColumnInfo var time: String = "",
-    @ColumnInfo var size: Float = 0f,
+    @ColumnInfo var size: Double = 0.0,
     @ColumnInfo var price: Float = 0f,
     @ColumnInfo var side: String = ""
 )
